+++ src/constants/colorVriant.ts
@@ -1,0 +1,3 @@
+export const Coloer = {
+    
+}
